requirements.txt
flask
pypdf
anthropic
python-dotenv
python-docx
